Query HTTP headers in a case-sensitive manner
[RFC 2616](http://www.w3.org/Protocols/rfc2616/rfc2616-sec4.html#sec4)
says that HTTP headers are meant to be case-insensitive. protocol-http handles headers by lower-casing their names internally when storing the headers.

However, when retrieving headers, protocol-http does not lowercase the header key. It stores the header lowercases it in a case-insensitive manner, and encodes the headers correctly for HTTP2 encoding, but does not correctly query headers in a case-insensitive manner. This means that querying for the key in the same casing you added it to the headers fails to return the value.